# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Agenda Risetec",
        "summary": "Routes of app",
        "description": "API under development",
        "termsOfService": "http://example.com/terms/",
        "contact": {
            "name": "Jhonattan Rocha da Silva",
            "url": "http://www.example.com/contact/",
            "email": "jhonattab246rocha@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "identifier": "MIT"
        },
        "version": "0.0.1"
    },
    "paths": {
        "/crud/generic": {
            "post": {
                "summary": "Generic Create",
                "operationId": "generic_create_crud_generic_post",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "model",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Model"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenericCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "summary": "Generic Reads",
                "operationId": "generic_reads_crud_generic_get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Model"
                        }
                    },
                    {
                        "name": "filters",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Filters"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/GenericCreate"
                                    },
                                    "title": "Response Generic Reads Crud Generic Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/crud/generic/{generic_id}": {
            "get": {
                "summary": "Generic Read",
                "operationId": "generic_read_crud_generic__generic_id__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "generic_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Generic Id"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "",
                            "title": "Model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "summary": "Generic Update",
                "operationId": "generic_update_crud_generic__generic_id__put",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "generic_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Generic Id"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Model"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/GenericCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "summary": "Generic Delete",
                "operationId": "generic_delete_crud_generic__generic_id__delete",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "generic_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Generic Id"
                        }
                    },
                    {
                        "name": "model",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Model"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenericCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/crud/log/": {
            "post": {
                "summary": "Create Log",
                "operationId": "create_log_crud_log__post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoggerBase"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LoggerCreate"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ]
            }
        },
        "/crud/logs/": {
            "get": {
                "summary": "Read Logs",
                "operationId": "read_logs_crud_logs__get",
                "security": [
                    {
                        "OAuth2PasswordBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "filters",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "title": "Filters"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Skip"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Logger"
                                    },
                                    "title": "Response Read Logs Crud Logs  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Event": {
                "properties": {
                    "title": {
                        "type": "string",
                        "title": "Title"
                    },
                    "description": {
                        "type": "string",
                        "title": "Description",
                        "default": ""
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Date"
                    },
                    "isAllDay": {
                        "type": "boolean",
                        "title": "Isallday"
                    },
                    "startTime": {
                        "type": "string",
                        "title": "Starttime"
                    },
                    "endTime": {
                        "type": "string",
                        "title": "Endtime"
                    },
                    "color": {
                        "type": "string",
                        "title": "Color"
                    },
                    "calendar_id": {
                        "type": "integer",
                        "title": "Calendar Id"
                    },
                    "user_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "User Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "title",
                    "date",
                    "isAllDay",
                    "startTime",
                    "endTime",
                    "color",
                    "calendar_id",
                    "id"
                ],
                "title": "Event"
            },
...
            "GenericCreate": {
                "properties": {
                    "values": {
                        "additionalProperties": true,
                        "type": "object",
                        "title": "Values",
                        "default": {}
                    },
                    "model": {
                        "type": "string",
                        "title": "Model",
                        "default": ""
                    }
                },
                "type": "object",
                "title": "GenericCreate"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Logger": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "entity": {
                        "type": "string",
                        "title": "Entity"
                    },
                    "data": {
                        "type": "string",
                        "title": "Data",
                        "default": "2025-06-20T08:20:52.064449"
                    },
                    "action": {
                        "type": "string",
                        "title": "Action"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "user": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/User"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "entity",
                    "action",
                    "id",
                    "user"
                ],
                "title": "Logger"
            },
            "LoggerBase": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "entity": {
                        "type": "string",
                        "title": "Entity"
                    },
                    "data": {
                        "type": "string",
                        "title": "Data",
                        "default": "2025-06-20T08:20:52.064449"
                    },
                    "action": {
                        "type": "string",
                        "title": "Action"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "entity",
                    "action"
                ],
                "title": "LoggerBase"
            },
            "LoggerCreate": {
                "properties": {
                    "user_id": {
                        "type": "integer",
                        "title": "User Id"
                    },
                    "entity": {
                        "type": "string",
                        "title": "Entity"
                    },
                    "data": {
                        "type": "string",
                        "title": "Data",
                        "default": "2025-06-20T08:20:52.064449"
                    },
                    "action": {
                        "type": "string",
                        "title": "Action"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "user_id",
                    "entity",
                    "action",
                    "id"
                ],
                "title": "LoggerCreate"
            },
            "Permissions": {
                "properties": {
                    "entity_name": {
                        "type": "string",
                        "title": "Entity Name"
                    },
                    "can_view": {
                        "type": "boolean",
                        "title": "Can View"
                    },
                    "can_delete": {
                        "type": "boolean",
                        "title": "Can Delete"
                    },
                    "can_update": {
                        "type": "boolean",
                        "title": "Can Update"
                    },
                    "can_create": {
                        "type": "boolean",
                        "title": "Can Create"
                    },
                    "profile_id": {
                        "type": "integer",
                        "title": "Profile Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "entity_name",
                    "can_view",
                    "can_delete",
                    "can_update",
                    "can_create",
                    "profile_id",
                    "id"
                ],
                "title": "Permissions"
            },
...
            "User": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "email": {
                        "type": "string",
                        "title": "Email"
                    },
                    "profile_id": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Profile Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "profile": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/UserProfile"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "events": {
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/Event"
                                },
                                {
                                    "type": "null"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Events"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "email",
                    "id",
                    "profile",
                    "events"
                ],
                "title": "User"
            },
...
            "UserProfile": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    },
                    "permissions": {
                        "items": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/Permissions"
                                },
                                {
                                    "type": "null"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Permissions",
                        "default": []
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "id"
                ],
                "title": "UserProfile"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "OAuth2PasswordBearer": {
                "type": "oauth2",
                "flows": {
                    "password": {
                        "scopes": {},
                        "tokenUrl": "/crud/token"
                    }
                }
            }
        }
    }
}